// frontend/src/app/core/enums/input-share-document.enum.ts
export enum EnumShareDocumentType {
    PHONE = 'phone',
    EMAIL = 'email',
    CONTACT = 'contact',
}

export enum EnumShareDocumentLabel {
    PHONE = 'Celular',
    EMAIL = 'Correo',
    CONTACT = 'Contacto',
}

export enum EnumShareDocumentActions {
    ADD = 'add',
    ATTACH = 'attach',
    LIST = 'list',
}

export enum EnumCheckBox {
    WHATSAPP = 'WhatsApp',
    SMS = 'SMS',
    QR = 'Incluir QR',
}

export enum EnumOptionShared {
    EMAIL = '1',
    SMS = '2',
    WHATSAPP = '3',
    CONTACT = '4',
}

export enum EnumShareDocumentOther {
    VIEW_ALL = 'Ver todos',
    VIEW_ATTACHMENTS = 'Ver listado carga masiva',
    LIST_MAX = 2,
    INITIAL_COUNTRY = 'co',
    EMPTY = ''
}